// src/api/models/FabricaResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type FabricaResponseDto = {
  /**
   * @type string
   */
  fabricaId: string
  /**
   * @type string, date-time
   */
  date: Date
  /**
   * @type boolean
   */
  principal: boolean
  /**
   * @type boolean
   */
  checkPoint: boolean
  /**
   * @type string
   */
  autor: string
  /**
   * @type string
   */
  fabricaPai: string | null
}